server/src/test/java/com/kvstore/raft/RaftNodeTest.java

package com.kvstore.raft;

import com.kvstore.storage.KVStorage;
import org.junit.Before;
import org.junit.Test;
import static org.junit.Assert.*;
import static org.mockito.Mockito.*;

import java.util.Arrays;

public class RaftNodeTest {
    private RaftNode raftNode;
    private KVStorage mockStorage;

    @Before
    public void setUp() {
        mockStorage = mock(KVStorage.class);
        raftNode = new RaftNode("node1", Arrays.asList("node2", "node3"), mockStorage);
    }

    @Test
    public void testInitialState() {
        assertFalse(raftNode.isLeader());
    }

    @Test
    public void testProposeEntry() {
        // This test is simplified and doesn't account for the asynchronous nature of proposeEntry
        // Todo: use a CountDownLatch or similar mechanism for more accurate testing
        raftNode.start(); // Assuming this method exists to initialize the RaftNode
        assertFalse(raftNode.proposeEntry("key", "value").join());
    }

    // Add more tests when implement more functionality
}